Implement Special Energy
Implemented Special Energy, which are handled in the game engine the
same way as resupply bases. The only difference between a resupply base
and a special energy station is the tile on the map.

On eligible stages (currently only 5) 3 of the resupply bases requested
will be upgraded to special energy stations, so make sure to account for
this when setting map parameters.