Randomize location of super energy bank
It's critical for this to be available, because if players opt to
collect super energy bases before engaging Mecha King Ghidorah and don't
start the battle before destruction of Tokyo is 50%, they'll be unable
to become Super Godzilla before confronting Bagan and be unable to
complete the game.

This could potentially be disabled in the future with a randomizer flag for masochistic players.